Update folder structure and imports
